XXX HACK XXX: minimally work around broken ^
Since StringScanner will successfully match /^b/ in "ab\n",
we introspect the regexp and use StringScanner's brilliant
method `#beginning_of_line?` to short-circuit if it shouldn't
match.  This still doesn't cover uses of ^ other than the
beginning of the regex, but that's the case that's most needed
in actual lexers.

See https://bugs.ruby-lang.org/issues/7092 for more info.