Improve diff lexer
The diff lexer was a bit buggy. It also did not work correctly
with email patches. This commit ensures correct patch highlighting.

Furthermore I removed the error token which caused problems on gitlab.
As a diff is rather trivial I think it is safe to remove it.

see: https://github.com/rumpelsepp/rugments/issues/10