Create lexer for the Julia programming language
Resolves #320

The Julia lexer here is now on par with that of Pygments. In some cases
it highlights differently than Pygments because there are more Types
(https://github.com/jneen/rouge/pull/331/files#diff-bab61e513a4a11b97616b453e4184909R40)
listed in this version of the lexer and some operators (ex. "::") are
treated a little different. Since Pygments doesn't differentiate string
interpolation, that wasn't prioritized for this first-pass port. It's
currently minimally supported by only highlighting the `$` where it
starts in a string.