Create lexer for the Pascal programming language
Added rules for a basic Pascal lexer. It does a reasonable job of
highlighting Pascal code. There is a bit of room for improvement
such as adding support for the different 'flavours' of Pascal i.e.
Free Pascal, Turbo Pascal etc. A Delphi lexer can easily be extended
from this lexer, however due to my lack of experience with Delphi I
chose not to include this in.