As Token::name returns only the last part of token's name, the terminal256 formatter shouldn't use it to cache escape sequences.
Actually, there is some conflicts like:
- Comment.Single and String.Single are known as :Single
- Other, Name.Other, String.Other, Number.Other as :Other
So the style (escape sequence) for the first token of a
given type will be reused for the others.

Eg, with:
<?php
echo 'bar';

The string bar (String.Single) is highlighted, by the
terminal256 formatter like the comment (Comment.Single)
above. And vice-versa if we invert these two instructions.