igorpro: allow dot in objectnames
An element of the structure s like s.objectName should not be
highlighted as a function if objectName matches a predefined function
name.
The easiest way is to allow the dot for objectnames. This allows rouge to
handle the whole structure element as an objectname.

i.e. in s.winName the winName should not be partly highlighted. But if a
winName stands alone in the code with a leading space char ist should be highlighted
as usual.