Improve single line comment formatting in files with CRLF
The rule for single line comments used $ to match newline but not capture it.
If a file is using CRLF this means the CR gets included in the comment value
but not the LF. This shows up as an extra line in the formatted HTML output
because one is inside the span and one is after.

Change to include anything up to the first CR or LF in the following languages:
- actionscript
- csharp
- groovy
- javascript