Add source code and changelog links to gemspec
This makes it easier for tools to programatically find the source
code and changelog for the gem - for example, [Dependabot](https://dependabot.com/)
uses this when bumping your dependencies automatically to provide
a link to the changleog and any GitHub releases.

This is in as-yet-undeployed docs for RubyGems, merged in a couple
of weeks ago from https://github.com/rubygems/rubygems/pull/1960.