Move keywords_type logic from hql.rb to sql.rb superclass
I added the distinction between keywords_type and keywords to
HQL which subclasses SQL. @dblessing makes the excellent suggestion to
move that up to the SQL class, which might benefit too.

This commit does that. That also means adding a list of type keywords
to the SQL class. I opted to take MySQL's list, which I'm more familiar
with than other dialects. Choice of dialect is an arbitrary choice
anyway.

This *mostly* adds keywords but a few keywords were moved from
keywords to keywords_type. SET is a tricky one, as it's also a
statement, but highlighting it as a type works for what's probably the
more prevalent occurrence.