Added Motorola 68k assembly lexer (Devpac syntax)
Added new lexer for Motorola 68000 assembly, for the Devpac syntax.
There are numerous assembler for the 68000 with slightly different
syntax. Devpac is the only one I know, but it could be used as a starting
point to implement the other variants.

It doesn't cover the "older" Devpac syntax where comments can be put
after the last operand, at the end of the line, without a separator (`;`
or `*`). This would require the lexer to know the semantics of the
language, and that was a bit too much for my first lexer :wink:

I started it from the NASM lexer, but it's a bit simpler.

I initially associated it with `*.asm` files but that conflicts with
NASM. I also associated it with the `text/x-asm` mime type, but I
removed it as I'm not sure if that makes sense (why this one, and not
NASM?)

As a result there are no tests, since there's nothing to test in terms
of file/mime association.