Fix quoted lists
As discussed in #458, the first element in a quoted list should not be treated
as if it were a function. This commit adds a new rule for lists that begin with
a quotation mark and moves the operator rule to avoid an early match. This does
not handle the case where a list is quoted by use of the `quote` special form.