Allow user to specify HTML tag in HTMLLinewise
The `Rouge::Formatters::HTMLLinewise` class provides a user with the ability to
wrap each line in a code block in a tag. Currently, that tag is hard-coded to be
`<div>`. Arguably the more semantically correct tag to use would be `<code>`,
but regardless of whether this is correct or not, lines that are wrapped in this
way display unpleasantly in RSS readers.

This commit allows a user to specify the tag to use by passing a value with the
key `:tag_name` to the class's initialiser. If no tag is specified, the class
uses `<div>` as the default (ie. it is consistent with the current behaviour).

Finally, to enhance the readability of the HTML, a new line is added at the end
of each line.