Improve lexer development documentation
Rouge does not have a single document explaining how a lexer should
be developed. The README contains a simplified example and there are
existing lexers that can be looked at for reference but these are
inconsistent and do not clearly explain some of the more difficult
elements (eg. disambiguating conflicting filename globs).

This commit adds a docs directory with a basic guide. The guide is
written in Markdown and so is viewable directly on GitHub but it is
intended for use with YARD and uses YARD tags for cross referencing. For
this reason, the commit also updates the `.yardopts` file. The README is
also updated to refer to this guide.