Fix untyped methods
The Objective-C lexer prepends a rule to the `:root` state that is
intended to match a method definition. Although the regular expression
indicates that the return type of the method may not match anything, it
was nevertheless passing the result to `RegexLexer#recurse` regardless
of whether it was nil or not.

If this occurred with an untyped instance method, the output would
insert an additional `-` in the output. Checking whether to call
`RegexLexer#recurse` on the basis of whether the value was non-nil fixed
the problem. This fixes #1076.