Fix whitespace tokenisation in Go lexer (#1122)
As discussed in #1097, the Go lexer tokenises whitespace using the
`Other` token. This causes issues when rendering in HTML the code parsed
by Rouge because the use of `Other` will result in the whitespace being
wrapped in separate HTML tags, which in turn causes all newlines are
collapsed (due to CSS rules that expect newlines not to be wrapped in
seperate HTML tags).

The `Other` token's intended use is described as follows:

> Token for data not matched by a parser (e.g. HTML markup in PHP code)

This is not a correct characterisation of whitespace in Go and is not
consistent with other lexers. This commit changes the token to `Text`,
the token generally used by other lexers. It fixes #1097.