Coerce state names into symbols rather than strings (#1138)
Lexers may be defined with the names of states given as `String`s rather
than `Symbol`s. To reduce memory usage, we should internally use
`Symbol`s and this coerces any `String` into a `Symbol`. It does not
handle the exception raised if the object doesn't respond to `#to_sym`.