Introduce :detectable? singleton method for lexers (#1149)
`Rouge::Guessers::Source` filters lexers, in part, by whether the lexer
provides a `:detect?` singleton method. Calling
`methods(false).include?(:detect?)` frequently is a significant memory
drain as it generates short-lived arrays (of varying size per lexer)
which are immediately swept away by the GC.

This commit provides a `:detectable?` singleton method that caches the
result of the expression on the first evaluation, and send the cached
result on subsequent calls from the same lexer class.