Add lexer and registered DSL keywords
This commit adds the `lexer` keyword to the lexer DSL. An abstract
implementation is defined in `Rouge::Lexer` and a concrete
implementation is defined in `Rouge::RegexLexer`. The intention is that
lexer authors can wrap their lexing logic in a block and pass it to the
`lexer` class method of their immediate superclass. This will avoid the
instantiation of objects used in the lexer's logic until an object of
the type of the lexer is initialised.

This commit also adds the `registered` keyword to the more specific
regular expression lexer DSL. The intention is that lexer authors can
wrap their registered keywords in a block and pass this to
`RegexLexer.registered`. This will create a memoised instance method
thet can be used in the lexing logic. This will avoid the instantiation
of registered words until they are required.