Fix backtracking issues, add string interpolation
The previous rules for strings used a regular expression that combined
different elements of the string. This could, in a pathological case,
cause Rouge to get stuck trying to lex the code. This problem can be
fixed by lexing a quoted string in a separate state to `:root`.

In addition to adding states for quoted string, this commit also adds
support for basic string interpolation. The basic code is lifted from
the Ruby lexer. This fixes #974.