Restore support for opts to Lexer.lex (#1178)
Version 3.4.0 of Rouge moved from passing `opts[:continue]` to the class
method `.lex` to calling the new class method `.continue_lex`. As
implemented, the change was breaking and could cause Rouge to crash. 
This was not intentional and this commit reverts the change.
Internally, Rouge still uses `.continue_lex` but `.lex` will function as it
did in version 3.3.0. A warning is displayed for uses with `$VERBOSE` set
to true.