Enable Rubocop ambiguity warnings
Rouge currently disables Rubocop's AmbiguousRegexpLiteral cop and its
Lint/AmbiguousOperator. Presumably this was because of the large number
of regular expressions used in Rouge that are defined using the `/.../`
literal syntax.

This commit prepends all regular expressions of that form with `%r`. It
also tidies up a few uses of the `*%w()` construct that are present in
the codebase. This allows these warnings to be turned on without causing
any issues.

It should be noted that having the warnings disabled is not a problem in
and of itself. However, these constructs also raise warnings when
running Ruby with the `-w` switch. Getting Rouge to a point where it can
be run with the `-w` switch will allow for a safer development
environment.