BPF lexer
The BPF bytecode is used in Linux (in its current syntax since v3.15) to
extend the kernel at runtime.  Clang includes a backend to compile from C.

The particular syntax supported with this commit is that of LLVM's
disassembler, which as far as I can see, is the most common syntax used to
display BPF bytecode snippets.

I haven't specified filenames or mimetypes because there isn't a clear
consensus yet.  Most of the time, BPF snippets are directly posted as examples
in larger texts (blog posts, emails, etc.).  When that is not the case, they
sometimes have a .b extension, other times an Assembly extension.

I'm adding support for BPF in Rouge to be able to use in Jekyll-based blogs.
You can ping me if reviews are needed for future improvements.

Signed-off-by: Paul Chaignon <paul.chaignon@orange.com>