Rename class lookup method, support block passing
The method `Lexer.lookup_fancy` returns a Lexer subclass while the
method `Lexer.find_fancy` returns an instance of the Lexer subclass.
This is not evident from the name and so this commit renames
`lookup_fancy` to `find_fancy_class`.

Separately, the commit also adds support to consumers passing a block to
`find_fancy`. It is not envisaged that block passing will be common.

Finally, test are added.