Support promoted data constructors in Haskell lexer (#1027)
This commit adds support for promoted data constructors to the Haskell
lexer. Unfortunately there is a bit of ambiguity with char literals.
This commit errs on the side of char literks. The data constructor `A'`
would be `'A'` in promoted form but the lexer will highlight this as a
char literal.