Fix invalid parenthesis state in PowerShell lexer (#1222)
The update to the Shell lexer made as a result of #1216 should have
included an update to the PowerShell lexer. This is because the Shell
lexer renamed the `:paren` state. Since the PowerShell lexer relies on
this state, it needed to be updated to refer to the new state. This
commit does that.