Avoid unnecessary String duplication in HTML formatter (#1244)
The use of `String#gsub` in `Rouge::Formatters::HTML#span` will create a
copy of every string passed to it irrespective of whether a substitution
occurs. This can lead to unnecessary memory consumption. This commit
uses a private helper to substitute values conditionally and invokes
`String#gsub` only if the string matches the respective pattern.