Add support for transliteration in Perl lexer
Add support for tr/// and y/// by hitching along with s/// from
substitutions.  Also reduce the number of substitution rules by
taking an arbitrary delimiter, as with match below, which has the
benefit of matching more substitutions correctly.

Also move __DATA__ down to live with __END__, because it also
signals the end of parsing.