Disable RuboCop's AmbiguousRegexpLiteral cop
RuboCop checks for ambiguous regular expression literals by default. The
concern is avoiding situations where it is ambiguous whether the
argument being passed is a regular expression literal or a call to the
`/` method defined on the relevant object. The risk of this occurring in
Rouge is considered to be insubstantial and offset by the impact on the
clarity of Rouge's DSL. This commit disables the cop.