Add improvements (eg. transliteration) to Perl lexer (#1250)
This commit adds support for `tr///` and `y///` by hitching along with
`s///` in the substitution rules. It also reduces the number of
substitution rules necessary by taking an arbitrary delimiter (this has
the added benefit of matching more substitutions correctly).

Finally, the commit moves `__DATA__` down to live with `__END__`; a more
appropriate location as it also signals the end of parsing.