Add HTMLLineTable formatter (#1211)
This commit adds as a new formatter, HTMLLineTable, a hybrid of
HTMLTable and HTMLLineWise. Like HTMLLinewise, this formatter generates
a new `<tr>` element for every line of highlighted code. However unlike
HTMLLinewise, this formatter instead assigns a configurable id attribute
to each `<tr>` (corresponding to the current line number) and a
configurable classname attribute.