Update Ada lexer after review by @pyrmont.
The rule entering state :libunit_name was incorrectly classifying
whitespace as Keyword::Namespace in `limited with`. Add a spec to make
sure this is marked as Text.

In the :libunit_name state, detect keywords that shouldn't be there and
use those to recover quickly from the incorrectly entered state.

This recovery is more robust that the previous ping-pong with
:libunit_sep, so merge that state into libunit_name for simplicity.

Add specs to make sure we're recovering well.