Remove italics from preprocessor style rules
Three of the default themes that come with Rouge italicise text
tokenised with the `Comment::Preproc` token. This is consistent with the
styling applied to other tokens within the `Comment` group but is not
appropriate for a token that is typically used with code that can
include characters like `/` (the `Comment::Preproc` token is the
recommended token for PHP and ERB, for example).

This commit removes the italics attribute from the themes Magritte,
Thankful Eyes and Tulip. The styles are otherwise left unchanged.