Remove italics from preprocessor style rules (#1264)
Three of the default themes that come with Rouge italicise text
tokenised with the `Comment::Preproc` token. The `Comment::Preproc`
token is the recommended token for the tags in PHP and ERB that mark the
beginning of code in those languages.

Using italics is consistent with the styling applied to other tokens
within the `Comment` group but is not appropriate for a token that is 
typically used with code that can include characters like `|` (which
could be mistaken for `/` when italicised). 

This commit removes the italics attribute for text tokenised as
`Comment::Preproc` in the themes Magritte, Thankful Eyes and Tulip. The
styles are otherwise left unchanged.