Update single-char matches
The markup stuff has been removed, as it currently doesn't work (the <
and > characters were being dropped, which is not desired).

Additionally, applying some requests eventually broke the block
attributes, so a new solution to get these highlighted correctly needs
to be found.

If I can find a solution to the former, I can most likely also resolve
the latter, while reducing the codebase.