Add disambiguation for C++ header files (#1269)
C++ header files can use the `.h` extension. By default, Rouge
highlights these files with the C lexer. This can lead to incorrect
highlighting because certain keywords in C++ are not reserved in C. A
similar situation can occur with Objective-C files but Rouge provides a
disambiguation for Objective-C files to avoid this problem. This commit
adds a similar disambiguation for C++.