Reduce memory usage with a pre-generated cache
Rouge's design (and the fact this is Ruby) leads to a large amount of
memory usage. This usage is unnecessary if the language of the file to
be lexed is known as the vast majority of memory used is for object
allocation of lexers that are not going to be used.

This commit uses a caching mechanism to load a pre-generated cache of
details about the avialable lexers and thereby avoids instantiating
lexers that are not required. It falls back to instantiating all lexers
if guessing is used. A consumer can opt-out of the caching behaviour
entirely by requiring `rouge/all` rather than `rouge`.