[bbcbasic] fix `*` operators being misidentified as inline commands
This is achieved to inserting `goto :no_further_imperatives` every time we
identify a component of an expression, or reach a state where we expect an
expression next. Because these will not be balanced by the number of times we
return to :root, we can no longer use a state stack. This means that instead
of being able to do `mixin :expression` from state `:assembly2`, we need to
write out all the same rules again, but without the `goto`s.