[bbcbasic] use a different approach for detecting CLI commands
Using a pair of states just wan't working - there are just too many exceptions
to the rules to handle cleanly. So instead, here we take the approach of only
permitting `*` to be lexed as a command introducer if it is at the start of a
line, or follows `:` or one of a restrictive set of other keywords.

Unfortunately, it appears that the lookbehind regexp syntax `(?<=)` doesn't
work here, so we have to lex the string including the lead up to the `*` as a
multi-token rule. This is fine in all cases except when the first line of the
file is a `*` command, but hopefully this will be rare enough that nobody
cares.