SQL/HQL: move interpretation of double-quoted string to HQL
As noticed by @vidarh, this better represents the existence of many
different interpretations in different SQL dialects.

The suggestion in the code review was to maybe make this an optional
argument (e.g. `dialect`) to the lexer, but since we only have two
dialects at the moment, I feel it's clearer to stick with subclassing
as a way to define dialects for now.