SQL: move SET and COLLATE to keywords
Both of these are (part of) type specifications as well, and so
we have to choose where to put them. In  f3311f1ee814d04, I let
the interpretation as a type always take precedence.

This commit applies a subjective rule-of-least-suprise instead:

 - Most people will run into COLLATE as a function, not as part
   of a unicode collation rule for a column
 - Most people will run into SET as part of an UPDATE statement,
   not as part of a CHARACTER SET specification

I chose not to move YEAR and TIMESTAMP from that same commit.
They are functions as well as types, but their use as a function
is so closely related to the type that there's no confusion to be
feared.

I didn't move NATIONAL (from NATIONAL CHAR), NCHAR and PRECISION. They
were moved to be types in the same commit, but that's their only
interpretation.