Add HQL lexer and add types to SQL lexer (#880)
This commit adds a lexer for Hive Query Language (a dialect of SQL). 

In addition, it updates the SQL lexer to move certain keywords to a
types category and tokenises these keywords with the `Name::Builtin`
token (this is consistent with the approach of Pygments and Chroma).

For keywords that can be used both in a type specification and as a
function (e.g. `COLLATE`), the categorisation was based on the most
common use of the keyword.