Change method of saving MatLab built-in keywords (#1300)
To avoid instantiating a large number of strings, Rouge currently saves
the list of built-in keywords for MatLab to a file that is only read
when the MatLab lexer is in use. The keywords that are in this file are
generated by a Rake task that extracts the keywords from the official
MatLab documentation.

Since the time that the Rake task was created, MathWorks (the creators
of MatLab) have changed the way this page works such that it requires
JavaScript to load. This breaks the Rake task.

As an alternative, this commit includes the list of built-in keywords in
a YAML file. This is consistent with some of the other lexers that have
a large number of keywords (e.g. Apache lexer). It will, unfortunately,
require copying and pasting for future updates.