Fix escaping/interpolating in string and charlist literals in Elixir lexer (#1308)
The Elixir lexer does not properly handle escaping and interpolating in
string (`"foo"`) and charlist (`'foo'`) literals.

This PR adds general escaping rules for use in strings as well as adding
support for escaping and interpolating to charlist literals. It also
rationalises the states used for string literals. Examples to confirm
correct behaviour have been added to the visual sample.