Improve support for Gradle plugin names in Kotlin lexer (#1323)
The contents of Gradle's Kotlin DSL is essentially Kotlin code and
Rouge's Kotlin lexer is used for lexing these files. However, since
Gradle plugins can include dashes, these were not being lexed correctly.

This commit permits dashes in identifiers. This will mean syntactically
incorrect Kotlin is not highlighted with errors but for the sake of
simplicity this is considered an appropriate trade-off.