Fix pattern for values incorporating backslashes in JSON lexer (#1331)
It is possible for a JSON value that incorporates backslashes and colons
to incorrectly match a pattern in the `:name` state. In certain
situations, this can cause the lexing of a JSON file to hang indefinitely.

This commit adds `\\` to the excluded character class in the `:name`
state. This causes such strings to instead be matched in the `:value` state.