Support disassembler output in BPF lexer (#1346)
This commit updates the BPF lexer to support several constructs from the
two main disassemblers, bpftool and objdump:

- `w4` syntax for registers used in 32bit ALU instructions;
- `map[id:671]` to refer to pointer to the map with id 671;
- call `bpf_skb_load_bytes#7684912` to refer to a call to helper
  `bpf_skb_load_bytes`;
- line numbers;
- `(bf)` and `61 12 04 00 00 00 00 00` for the hexadecimal encoding of
  the opcode and instruction respectively.

Basic support is added for the hexadecimal output (Generic group) so
that it doesn't break highlighting. By supporting this, it's easier for
users to directly highlight the output of bpftool and objdump.