Fix comment parsing in Console lexer (#1379)
The Console lexer would not correctly parse comments if only the
`comments` option was enabled. This was because of a bug in the way the
prompt's regular expression pattern worked. This commit fixes that bug
and adds a series of tests to confirm the behaviour of the lexer is as
intended.