Fix parsing of 'false' as Boolean option value (#1382)
Rouge allows for options to be passed to a lexer's constructor method.
The options are parsed from strings and the string 'false' was not being
converted to the Boolean value `false`. This commit fixes that and also
adds a rudimentary series of tests to ensure that Boolean options are
being properly parsed.