Add Objective-C++ lexer (#1378)
This commit adds a lexer for Objective-C++.

Since Ruby does not support multiple inheritance, the Objective-C rules
previously defined in the Objective-C lexer are factored out into a
common module that can be used by both lexers. Accordingly, this commit
also updates the Objective-C lexer to use this module.

As the common module uses memoised methods (e.g. `at_keywords`) that are
referred to within rules included in the common module, lexers that use
the common module need to use `Object#extend` rather than
`Module#include`. Consequently, the memoised methods are not prefixed
with `self` as would typically be the case. This is different to other
factored-out modules intended for reuse in lexers (e.g. the
`TypescriptCommon` module).