Support optional props in TypeScript lexer (#1393)
TypeScript allows a user to mark a property as a optional when defining
an interface by appending a question mark to the property name. Rouge's
TypeScript lexer doesn't currently support this because the lexer merely
uses the states defined in the JavaScript lexer.

This commit prepends a rule to the `statement` state that will match a
property name ending with a `?`. It is noted that this is more
permissive than TypeScript's syntax.