Fix memoisation of Lexer.detectable? (#1425)
The current implementation of `Lexer.detectable?` uses the `||=` idiom
to memoise the value returned from checking the existence of a `detect?`
method. However, since that value may be `false` it means that lexers
without the method will always perform the check.

This commit instead first checks with `defined?` to see if the memoised
value has been set. If so, it is returned. Otherwise the check for
`detect?` is performed. That second check is performed using
`Object#singleton_methods` rather than `Object#methods` to better
communicate intent.