Fix symbol lexing in Scala lexer (#1438)
The Scala lexer currently matches symbols by ensuring that the last
character in the match is not a `'`.

The problem with this approach is that if the symbol occurs within a
larger structure (e.g. parentheses), the first character after the
symbol (e.g. the closing parenthesis) is matched as part of the symbol.

This commit uses a negative lookahead to avoid this problem.