Fix fenced code blocks in Markdown lexer (#1442)
Many flavours of Markdown permit fenced code blocks. These blocks may
begin with three or more ` or ~. The current implementation only permits
three. This commit fixes that.

The commit also fixes a bug in code blocks that would match the closing
fence even if it occurred within code in the block.