Fix brackets in links in Markdown lexer (#1445)
Brackets (with backticks) inside link text was not being highlighted
correctly in the Markdown lexer. This commit causes the lexer to search
for a closing bracket that is followed by either an opening bracket
(indicating this is a reference-style link) or an opening parenthesis
(indicating this is an inline link).