Revert Rubocop splat expansion cop (#1461)
This commit reverses the change made in #1451 and changes the name of
the cop back to `UnneededSplatExpansion`.

The reason to do this is that Rouge supports versions of Ruby from 2.0
and this prevents upgrading our development dependencies to a newer
version of RuboCop (RuboCop ended support for versions prior to v2.3
with RuboCop v0.50). The renamed cop was not introduced until RuboCop 
v0.76.

Dropping support for older versions of Ruby may be implemented as part
of Rouge v4.0 but it won't occur as part of v3.0.