Separate ? from other operators in PHP lexer (#1478)
The PHP lexer has a rule that can match one or more characters as
operators. The problem is that if an operator occurs immediately before
the `?>` that closes a block of PHP, the `?` will match as an operator.
This is possible with the `:` operator. This commit fixes the problem
by splitting `?` out into its own rule.