Add option to tokenise error messages in Console lexer (#1498)
The Console lexer is intended for use in lexing text that would be
expected in a console. As a convenience, it allows an end-user to
specify prompts and comment-demarcating characters.

As a similar convenience, this commit adds an `error` list option. By
default the value is `nil` but if specified by the end-user, any line
(not already identified as a snip, prompt or comment) which begins with
one of the strings in the option will be tokenised as `Generic::Error`.

To assist in future review and maintenance, this commit also
reorganises the methods in the class into alphabetical order.