Use plain Ruby files for built-in keyword lists (#1418)
Rouge has a number of lexers that include a list of the names of
built-in functions for more accurate tokenisation. Some of these lists
consist of YAML files that need to be computed at runtime. Rather than
take this approach, this commit ensures that the lists are plain Ruby
files that can be loaded by the respective lexers.