Use generic string states in Python lexer (#1477)
Python allows for a variety of string literals (formatted, raw,
unicode) as well as byte literals. In addition, strings can be
delimited by `'`, `"`, `'''` and `"""`. At present, the Python lexer
contains multiple states to handle the supported combination. This
approach is duplicative, error-prone and doesn't scale.

This commit takes a different approach. A `StringRegister` class is
added to the Python lexer and used to record the stack of string
literals currently being lexed. This makes it possible to implement
generic string states and then apply the appropriate tokens based on
the state of the register.