Add Rake task to generate keywords for LLVM lexer (#1505)
The list of keywords that is recognised in the current LLVM lexer is
not as complete as in some other syntax highlighting libraries (such as 
Pygments). This commit adds a Rake task that checks a source file in
the public LLVM repo and generates a keyword file that is loaded on
request when the LLVM lexer is used.