Update keyword-generation Rake tasks (#1500)
A number of the Rake tasks that generate lists of built-in keywords
for particular languages have bugs. These range from actual errors (as
in the case of the Lua task) to deprecation warnings (in the case of
tasks using `Kernel#open` to open URIs).

This commit fixes these bugs and includes freshly generated lists of
the built-in keywords that have changed.